+++ semantic-conventions/CHANGELOG.md
@@ -5,7 +5,7 @@ Please update the changelog as part of any significant pull request.
 ## Unreleased
 
 - Support `frozen` stability level and change default stability level from `stable` to `experimental`
-  ([#TODO](https://github.com/open-telemetry/build-tools/pull/TODO))
+  ([#170](https://github.com/open-telemetry/build-tools/pull/170))
 
 ## v0.18.0